Improve appmap testing

+++ spec/rails_spec_helper.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 def wait_for_container(app_name)
   start_time = Time.now
-  until (`docker ps -q -f name=#{app_name} -f health=healthy`.strip) != ''
+  until (`docker-compose ps -q --filter health=healthy #{app_name}`.strip) != ''
     elapsed = Time.now - start_time
     raise "Timeout waiting for container #{app_name} to be ready" if elapsed > 10
 
@@ -11,20 +11,24 @@ def wait_for_container(app_name)
   end
 end
 
-shared_examples_for 'Rails app pg database' do
+shared_context 'Rails app pg database' do
   before(:all) do
-    cmd = 'docker-compose up -d pg'
-    system cmd, chdir: 'spec/fixtures/rails_users_app' or raise "Command failed: #{cmd}"
-    wait_for_container 'rails_users_app_pg'
+    raise "you must set @fixure_dir" unless @fixture_dir
+    
+    Dir.chdir @fixture_dir do 
+      cmd = 'docker-compose up -d pg'
+      system cmd or raise "Command failed: #{cmd}"
+      wait_for_container 'pg'
 
-    cmd = 'docker-compose run --rm app ./create_app'
-    system cmd, chdir: 'spec/fixtures/rails_users_app' or raise "Command failed: #{cmd}"
+      cmd = 'docker-compose run --rm app ./create_app'
+      system cmd or raise "Command failed: #{cmd}"
+    end
   end
 
   after(:all) do
     if ENV['NOKILL'] != 'true'
       cmd = 'docker-compose down -v'
-      system cmd, chdir: 'spec/fixtures/rails_users_app' or raise "Command failed: #{cmd}"
+      system cmd, chdir: @fixture_dir or raise "Command failed: #{cmd}"
     end
   end
 end
